# openapi.yaml
openapi: 3.0.0
info:
  title: Pet Adoption System API
  description: API for user registration, login, adopter preferences, shelter pet
    management, and recommendations in the Pet Adoption System.
  version: 5.0.0
  contact:
    name: PRJ666 - Group 1 - 'Pet Adoption System' Team
    email: support@petadoption.com
servers:
  - url: ${openapi.petAdoptionSystem.base-path}
    description: Configurable base path for the API
...
paths:
...
  /pet/add-profile:
    post:
      tags:
        - pet
      summary: Add a new pet profile to the system
      security:
        - BearerAuth: [ ]
      description: Add a new pet profile to the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Buddy
                comments:
                  type: string
                  example: Friendly and playful
      responses:
        '201':
          description: Pet added successfully
          $ref: '#/components/responses/GenericApiResponse'
        '405':
          description: Invalid input
          $ref: '#/components/responses/GenericApiResponse'
  /pet/upload-image:
    post:
      tags:
        - pet
      summary: Upload an image for a pet
      security:
        - BearerAuth: [ ]
      description: Uploads an image file for a specific pet.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
              required:
                - image
      responses:
        '200':
          description: Image uploaded
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      payload:
                        type: string
                        example: /images/someImage.jpg
        '400':
          description: Invalid request or missing parameters
          $ref: '#/components/responses/GenericApiResponse'
  /pet/get-profile:
    get:
      tags:
        - pet
      summary: Retrieve all pet for the current user (shelter)
      security:
        - BearerAuth: [ ]
      description: Fetches all pet profile details for the current user (shelter).
      responses:
        '200':
          description: Profile details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      payload:
                        type: array
                        items:
                          $ref: '#/components/schemas/Pet'
        '401':
          description: Unauthorized access
          $ref: '#/components/responses/GenericApiResponse'
  /pet/get-profile/{petId}:
    get:
      tags:
        - pet
      summary: Retrieve pet profile information
      security:
        - BearerAuth: [ ]
      description: Fetches the pet profile details.
      parameters:
        - name: petId
          in: path
          description: The unique ID of the pet profile to delete.
          required: true
          schema:
            type: string
            example: "12abd45f...."
      responses:
        '200':
          description: Profile details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      payload:
                        $ref: '#/components/schemas/Pet'
        '401':
          description: Unauthorized access
          $ref: '#/components/responses/GenericApiResponse'
  /pet/update-profile:
    put:
      tags:
        - pet
      summary: Update pet profile
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Buddy
                comments:
                  type: string
                  example: Updated comments about Buddy
      responses:
        '200':
          description: Pet profile updated successfully
          $ref: '#/components/responses/GenericApiResponse'
  /pet/delete-profile/{petId}:
    delete:
      tags:
        - pet
      summary: Deletes a pet profile
      security:
        - BearerAuth: [ ]
      description: Deletes the pet profile with the specified ID.
      parameters:
        - name: petId
          in: path
          description: The unique ID of the pet profile to delete.
          required: true
          schema:
            type: string
            example: "12abd45f...."
      responses:
        '204':
          description: Pet profile successfully deleted.
          $ref: '#/components/responses/GenericApiResponse'
        '400':
          description: Invalid pet profile value
          $ref: '#/components/responses/GenericApiResponse'
        '401':
          description: Unauthorized access.
          $ref: '#/components/responses/GenericApiResponse'
        '403':
          description: Forbidden. User does not have permission to delete this profile.
          $ref: '#/components/responses/GenericApiResponse'
        '404':
          description: Pet not found.
          $ref: '#/components/responses/GenericApiResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # Optional, but helps to specify the token type
  responses:
    GenericApiResponse:
      description: Generic API response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
...
    Pet:
      properties:
        petId:
          type: string
          example: "0adf23ff..."
        shelterId:
          type: string
          example: "0adf23ff..."
        petName:
          type: string
          example: "Snooppy"
        images:
          type: array
          items:
            type: string
          example: ["http://some.site/images/image.jpg"]
        attributes:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
            description: >
              A map where each key is the name of an attribute group, and the value is a list of
              selected preferences within that group.
            example:
              "gender": [ "male" ]
              "environment": [ "house", "condo" ]
      type: object
...